requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
